Feat: Create models

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/crediya/loan/r2dbc/entity/LoansEntity.java
@@ -25,7 +25,7 @@ public class LoansEntity extends AbstractAuditoriaEntity{
     @Column("monto")
     private BigDecimal amount;
     @Column("plazo")
-    private BigDecimal term;
+    private Integer term;
     @Column("identification")
     private String identification;
     @Column("id_estado")
